refactor: moves analytics screenview to SceneLifecycle to track unlocking the phone and not break things
