+++ Sources/Screens/Unlock/UnlockScreenViewModel.swift
@@ -20,10 +20,7 @@ struct UnlockScreenViewModel: BaseViewModel {
     }
 
     func didAppear() {
-        let screen = ScreenView(id: BiometricEnrollmentAnalyticsScreenID.unlockScreen.rawValue,
-                                screen: BiometricEnrollmentAnalyticsScreen.unlockScreen,
-                                titleKey: "one login unlock screen")
-        analyticsService.trackScreen(screen)
+
     }
     
     func didDismiss() {
