+++ Tests/UnitTests/Application/SceneLifecycleTests.swift
@@ -1,3 +1,4 @@
+import GDSAnalytics
 @testable import OneLogin
 import XCTest
 
@@ -56,7 +57,15 @@ extension SceneLifecycleTests {
     }
     
     func test_promptToUnlock() throws {
+        XCTAssertEqual(mockAnalyticsService.screensVisited.count, 0)
         sut.promptToUnlock()
+        XCTAssertEqual(mockAnalyticsService.screensVisited.count, 1)
+        let screen = ScreenView(id: BiometricEnrollmentAnalyticsScreenID.unlockScreen.rawValue,
+                                screen: BiometricEnrollmentAnalyticsScreen.unlockScreen,
+                                titleKey: "one login unlock screen")
+        XCTAssertEqual(mockAnalyticsService.screensVisited, [screen.name])
+        XCTAssertEqual(mockAnalyticsService.screenParamsLogged["title"], screen.parameters["title"])
+        XCTAssertEqual(mockAnalyticsService.screenParamsLogged["screen_id"], screen.parameters["screen_id"])
         XCTAssertTrue(mockWindowManager.hideUnlockWindowCalled)
     }
 }
